flask_graphql_project/requirements.txt
alembic==1.5.8
aniso8601==7.0.0
click==7.1.2
Flask==1.1.2
Flask-Classful==0.14.2
Flask-GraphQL==2.0.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
promise==2.3
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
SQLAlchemy==1.4.11
Werkzeug==1.0.1
